pyproject.toml
[project]
name = "mcp-part-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "logfire>=4.18.0",
    "openai-agents>=0.6.5",
    "python-dotenv>=1.2.1",
]
